rename logging to slog (structured log)
Structured logging should only be used by servers (e.g. shellrepeater),
and should not be used by the earthly command. This renames logging to
slog and gives a comment describing when to use slog.
