+++ cmd/shellrepeater/main.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/earthly/earthly/debugger/server"
-	"github.com/earthly/earthly/logging"
+	"github.com/earthly/earthly/slog"
 
 	"github.com/sirupsen/logrus"
 )
@@ -14,7 +14,7 @@ const addr = "0.0.0.0:8373"
 func main() {
 	logrus.SetLevel(logrus.DebugLevel)
 	ctx := context.Background()
-	log := logging.GetLogger(ctx).With("app", "shellrepeater")
+	log := slog.GetLogger(ctx).With("app", "shellrepeater")
 
 	x := server.NewServer(addr, log)
 	x.Start()
